+++ src/joyride/getting_started.cljs
@@ -4,13 +4,30 @@
             [clojure.string :as string]
             [joyride.config :as conf]
             [joyride.utils :as utils]
+            [joyride.when-contexts :as when-contexts]
             [promesa.core :as p]))
 
-(defn- path->uri [base-path sub-path]
-  (apply (.-joinPath vscode/Uri) (vscode/Uri.file base-path) sub-path))
+(defn update-script-contexts!
+  "Updates VS Code context variables based on current script file existence"
+  []
+  (p/let [user-activate-exists? (utils/path-or-uri-exists?+
+                                 (utils/path->uri (conf/user-abs-scripts-path) ["user_activate.cljs"]))
+          user-hello-exists? (utils/path-or-uri-exists?+
+                              (utils/path->uri (conf/user-abs-scripts-path) ["hello_joyride_user_script.cljs"]))
+          ws-scripts-path (conf/workspace-abs-scripts-path)
+          ws-activate-exists? (when ws-scripts-path
+                                (utils/path-or-uri-exists?+
+                                 (utils/path->uri ws-scripts-path ["workspace_activate.cljs"])))
+          ws-hello-exists? (when ws-scripts-path
+                             (utils/path-or-uri-exists?+
+                              (utils/path->uri ws-scripts-path ["hello_joyride_workspace_script.cljs"])))]
+    (when-contexts/set-context! ::when-contexts/joyride.userActivateScriptExists user-activate-exists?)
+    (when-contexts/set-context! ::when-contexts/joyride.userHelloScriptExists user-hello-exists?)
+    (when-contexts/set-context! ::when-contexts/joyride.workspaceActivateScriptExists ws-activate-exists?)
+    (when-contexts/set-context! ::when-contexts/joyride.workspaceHelloScriptExists ws-hello-exists?)))
 
 (defn- getting-started-content-uri [sub-path]
-  (path->uri (utils/extension-path) (concat ["assets" "getting-started-content"] sub-path)))
+  (utils/path->uri (utils/extension-path) (concat ["assets" "getting-started-content"] sub-path)))
 
 (defn- create-content-file+ [source-uri ^js destination-uri]
   (js/console.info "Creating " ^String (.-fsPath destination-uri))
@@ -19,6 +36,7 @@
                                                   path/dirname
                                                   vscode/Uri.file))
         (vscode/workspace.fs.copy source-uri destination-uri)
+        (update-script-contexts!)
         destination-uri))
 
 (defn- maybe-create-content+
@@ -31,47 +49,73 @@
 
 (defn maybe-create-user-content+ []
   (maybe-create-content+ (getting-started-content-uri ["user" "deps.edn"])
-                         (path->uri (conf/user-abs-joyride-path) ["deps.edn"]))
-  (p/let [user-activate-uri (path->uri (conf/user-abs-scripts-path) ["scripts" "user_activate.cljs"])
+                         (utils/path->uri (conf/user-abs-joyride-path) ["deps.edn"]))
+  (p/let [user-activate-uri (utils/path->uri (conf/user-abs-scripts-path) ["scripts" "user_activate.cljs"])
           user-activate-exists?+ (utils/path-or-uri-exists?+ user-activate-uri)]
     (when-not user-activate-exists?+
       (maybe-create-content+ (getting-started-content-uri ["user" "scripts" "user_activate.cljs"])
-                             (path->uri (conf/user-abs-scripts-path) ["user_activate.cljs"]))
+                             (utils/path->uri (conf/user-abs-scripts-path) ["user_activate.cljs"]))
       (maybe-create-content+ (getting-started-content-uri ["user" "scripts" "hello_joyride_user_script.cljs"])
-                             (path->uri (conf/user-abs-scripts-path) ["hello_joyride_user_script.cljs"]))
+                             (utils/path->uri (conf/user-abs-scripts-path) ["hello_joyride_user_script.cljs"]))
       (maybe-create-content+ (getting-started-content-uri ["user" "scripts" "hello_joyride_user_script.js"])
-                             (path->uri (conf/user-abs-scripts-path) ["hello_joyride_user_script.js"]))
+                             (utils/path->uri (conf/user-abs-scripts-path) ["hello_joyride_user_script.js"]))
       (maybe-create-content+ (getting-started-content-uri ["user" "src" "my_lib.cljs"])
-                             (path->uri (conf/user-abs-src-path) ["my_lib.cljs"])))))
+                             (utils/path->uri (conf/user-abs-src-path) ["my_lib.cljs"])))))
 
-(defn maybe-create-workspace-config+ [create-joyride-dir?]
-  (p/let [joyride-dir-exists?+ (utils/path-or-uri-exists?+ (path->uri (conf/workspace-abs-joyride-path) "."))]
-    (when (or joyride-dir-exists?+ create-joyride-dir?)
-      (p/let [deps-uri (path->uri (conf/workspace-abs-joyride-path) ["deps.edn"])
-              _created?+ (maybe-create-content+ (getting-started-content-uri ["workspace" "deps.edn"])
-                                                deps-uri)]
-        deps-uri))))
+(defn maybe-create-user-readme+ []
+  (maybe-create-content+ (getting-started-content-uri ["user" "README.md"])
+                         (utils/path->uri (conf/user-abs-joyride-path) ["README.md"])))
+
+(defn maybe-create-workspace-config+ []
+  (maybe-create-content+ (getting-started-content-uri ["workspace" "deps.edn"])
+                         (utils/path->uri (conf/workspace-abs-joyride-path) ["deps.edn"])))
+
+(defn maybe-create-user-config+ []
+  (maybe-create-content+ (getting-started-content-uri ["user" "deps.edn"])
+                         (utils/path->uri (conf/user-abs-joyride-path) ["deps.edn"])))
 
 (defn create-and-open-content-file+ [source destination]
-  (fn []
-    (p/-> (create-content-file+ source destination)
-          (vscode/workspace.openTextDocument)
-          (vscode/window.showTextDocument
-           #js {:preview false, :preserveFocus false}))))
+  (p/-> (create-content-file+ source destination)
+        (vscode/workspace.openTextDocument)
+        (vscode/window.showTextDocument
+         #js {:preview false, :preserveFocus false})))
 
 (defn maybe-create-and-open-content+ [source destination]
   (p/let [exists?+ (utils/path-or-uri-exists?+ destination)]
     (when-not exists?+
       (create-and-open-content-file+ source destination))))
 
-(defn maybe-create-workspace-activate-fn+ []
+(defn maybe-create-workspace-activate-script+ []
   (p/do
-    (maybe-create-workspace-config+ true)
-    (maybe-create-and-open-content+ (getting-started-content-uri ["workspace" "scripts" "workspace_activate.cljs"])
-                                    (path->uri (conf/workspace-abs-scripts-path) ["workspace_activate.cljs"]))))
+    (maybe-create-workspace-config+)
+    (maybe-create-and-open-content+
+     (getting-started-content-uri ["workspace" "scripts" "workspace_activate.cljs"])
+     (utils/path->uri (conf/workspace-abs-scripts-path) ["workspace_activate.cljs"]))))
 
-(defn maybe-create-workspace-hello-fn+ []
+(defn maybe-create-workspace-hello-script+ []
   (p/do
-    (maybe-create-workspace-config+ true)
-    (maybe-create-and-open-content+ (getting-started-content-uri ["workspace" "scripts" "hello_joyride_workspace_script.cljs"])
-                                    (path->uri (conf/workspace-abs-scripts-path) ["hello_joyride_workspace_script.cljs"]))))
+    (maybe-create-workspace-config+)
+    (maybe-create-and-open-content+
+     (getting-started-content-uri ["workspace" "scripts" "hello_joyride_workspace_script.cljs"])
+     (utils/path->uri (conf/workspace-abs-scripts-path) ["hello_joyride_workspace_script.cljs"]))))
+
+(defn maybe-create-user-activate-script+
+  []
+  (p/do
+    (maybe-create-user-config+)
+    (maybe-create-and-open-content+
+     (getting-started-content-uri ["user" "src" "joy_button.cljs"])
+     (utils/path->uri (conf/user-abs-src-path) ["joy_button.cljs"]))
+    (maybe-create-and-open-content+
+     (getting-started-content-uri ["user" "scripts" "user_activate.cljs"])
+     (utils/path->uri (conf/user-abs-scripts-path) ["user_activate.cljs"]))))
+
+(defn maybe-create-user-hello-script+
+  []
+  (p/do
+    (maybe-create-user-config+)
+    (maybe-create-content+ (getting-started-content-uri ["user" "scripts" "hello_joyride_user_script.js"])
+                           (utils/path->uri (conf/user-abs-scripts-path) ["hello_joyride_user_script.js"]))
+    (maybe-create-and-open-content+
+     (getting-started-content-uri ["user" "scripts" "hello_joyride_user_script.cljs"])
+     (utils/path->uri (conf/user-abs-scripts-path) ["hello_joyride_user_script.cljs"]))))