+++ vscode-test-runner/workspace-1/.joyride/src/integration_test/scripts_test.cljs
@@ -19,13 +19,15 @@
 
 (deftest-async run-a-user-script
   (testing "Runs a user script"
-    (p/let [result (vscode/commands.executeCommand "joyride.runUserScript" "hello_joyride_user_script.cljs")]
+    (p/let [_created (vscode/commands.executeCommand "joyride.createUserHelloScript")
+            result (vscode/commands.executeCommand "joyride.runUserScript" "hello_joyride_user_script.cljs")]
       (is (= "🎸"
              result)))))
 
 (deftest-async run-a-user-javascript-script
   (testing "Runs a user script written in JavaScript"
-    (p/let [result (vscode/commands.executeCommand "joyride.runUserScript" "hello_joyride_user_script.js")]
+    (p/let [_created (vscode/commands.executeCommand "joyride.createUserHelloScript")
+            result (vscode/commands.executeCommand "joyride.runUserScript" "hello_joyride_user_script.js")]
       (is (= "Hello World!"
              (.hello result))))))